requirements.txt
PyPDF2
langchain_google_genai
langchain
langchain_community
langchain_openai
biopython